fixes issues with probe utils moving
