+++ src/utils/probe-utils.js
@@ -1,8 +1,8 @@
 import produce from 'immer';
-import { fullBuildIDToDate } from '../patterns/utils/build-id-utils';
-import { nearestBelow } from '../../utils/stats';
-import { formatBuildIDToOnlyDate } from '../patterns/utils/formatters';
-import { groupBy } from '../../components/utils/transforms';
+import { fullBuildIDToDate } from '../app/patterns/utils/build-id-utils';
+import { nearestBelow } from './stats';
+import { formatBuildIDToOnlyDate } from '../app/patterns/utils/formatters';
+import { groupBy } from '../components/utils/transforms';
 
 export function sortByKey(key) {
   return (a, b) => {
